+++ client/src/main/java/client/scenes/OverviewCtrl.java
@@ -1,6 +1,7 @@
 package client.scenes;
 
 import client.Main;
+import client.UserConfig;
 import client.utils.ServerUtils;
 import client.utils.WebSocketUtils;
 import com.google.inject.Inject;
@@ -28,6 +29,7 @@ import javafx.scene.paint.Color;
 import javafx.scene.text.Text;
 import commons.Expense;
 import javafx.collections.ObservableList;
+import javafx.stage.Modality;
 
 
 import java.io.*;
@@ -41,7 +43,6 @@ import java.util.List;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.prefs.BackingStoreException;
-import static client.Main.prefs;
 import static client.Main.switchLocale;
 
 /**
@@ -97,6 +98,9 @@ public class OverviewCtrl implements Main.UpdatableUI {
     private Pane options;
     @FXML
     public AnchorPane ap;
+    private boolean admin;
+    private final UserConfig userConfig = new UserConfig();
+
 
 
     /**
@@ -117,15 +121,19 @@ public class OverviewCtrl implements Main.UpdatableUI {
      * Initializes the controller.
      */
     public void initialize() {
-        Image image = new Image(Objects.requireNonNull(getClass().getResource(prefs.get(SELECTED_IMAGE_KEY, "/client/misc/uk_flag.png"))).toExternalForm());
-        menuButtonView.setImage(image);
+        admin=false;
+        String lp = userConfig.getLanguageConfig();
+        if (lp.equals("en") || lp.equals("nl") || lp.equals("es")) {
+            Image image = new Image("/client/misc/" + lp +  "_flag.png");
+            menuButtonView.setImage(image);
+        }
+
         ap.setOnKeyPressed(event -> {
             if (event.getCode() == KeyCode.ESCAPE) {
                 backToStartScreen();
             }
         });
         expenseList = new ListView<>();
-        //TODO connect to the given server url when the initial startscreen is created
         webSocket.connect("ws://localhost:8080/websocket");
         webSocket.addEventListener((event)->{
             if(this.event==null||!this.event.getId().equals(event.getId())) return;
@@ -220,25 +228,33 @@ public class OverviewCtrl implements Main.UpdatableUI {
             label.setOnMouseClicked(mouseEvent -> {
                 if (mouseEvent.getClickCount() == 2) {
                     if (mouseEvent.getButton() == MouseButton.SECONDARY) {
-                        event.removeParticipant(contact);
-                        serverUtils.updateEvent(event);
-                        List<Expense> toDelete = new ArrayList<>();
-                        for(Expense expense : event.getExpenses()){
-                            if(expense.getPaidBy().equals(contact)) toDelete.add(expense);
-                            else if(expense.getInvolvedParticipants().contains(contact)){
-                                if(expense.getInvolvedParticipants().size()==1) toDelete.add(expense);
-                                else{
-                                    expense.getInvolvedParticipants().remove(contact);
-                                    serverUtils.updateExpense(event.getId(), expense);
+                        var alert = new Alert(Alert.AlertType.CONFIRMATION);
+                        alert.initModality(Modality.APPLICATION_MODAL);
+                        alert.setContentText("Are you sure you want to delete this participant?");
+                        alert.showAndWait().ifPresent((response)->{
+                            if(response == ButtonType.OK){
+                                event.removeParticipant(contact);
+                                serverUtils.updateEvent(event);
+                                List<Expense> toDelete = new ArrayList<>();
+                                for(Expense expense : event.getExpenses()){
+                                    if(expense.getPaidBy().equals(contact)) toDelete.add(expense);
+                                    else if(expense.getInvolvedParticipants().contains(contact)){
+                                        if(expense.getInvolvedParticipants().size()==1) toDelete.add(expense);
+                                        else{
+                                            expense.getInvolvedParticipants().remove(contact);
+                                            serverUtils.updateExpense(event.getId(), expense);
+                                        }
+                                    }
                                 }
+                                for(Expense expense1: toDelete){
+                                    serverUtils.deleteExpense(event.getId(), expense1);
+                                }
+                                serverUtils.deleteParticipant(contact);
+                                this.event = serverUtils.updateEvent(this.event);
+                                participantsDisplay();
                             }
-                        }
-                        for(Expense expense1: toDelete){
-                            serverUtils.deleteExpense(event.getId(), expense1);
-                        }
-                        serverUtils.deleteParticipant(contact);
-                        this.event = serverUtils.updateEvent(this.event);
-                        participantsDisplay();
+                        });
+
                     } else
                         addParticipant1(contact);
                 }
@@ -298,9 +314,9 @@ public class OverviewCtrl implements Main.UpdatableUI {
      * @param actionEvent
      */
     public void switchToEnglish(ActionEvent actionEvent) throws BackingStoreException {
+        userConfig.setLanguageConfig("en");
         switchLocale("messages", "en");
-        Image image = new Image(Objects.requireNonNull(getClass().getResource("/client/misc/uk_flag.png")).toExternalForm());
-        prefs.put(SELECTED_IMAGE_KEY, "/client/misc/uk_flag.png");
+        Image image = new Image(Objects.requireNonNull(getClass().getResource("/client/misc/en_flag.png")).toExternalForm());
         menuButtonView.setImage(image);
     }
 
@@ -309,9 +325,9 @@ public class OverviewCtrl implements Main.UpdatableUI {
      * @param actionEvent
      */
     public void switchToDutch(ActionEvent actionEvent) throws BackingStoreException {
+//        userConfig.setLanguageConfig("nl");
         switchLocale("messages", "nl");
         Image image = new Image(Objects.requireNonNull(getClass().getResource("/client/misc/nl_flag.png")).toExternalForm());
-        prefs.put(SELECTED_IMAGE_KEY, "/client/misc/nl_flag.png");
         menuButtonView.setImage(image);
     }
 
@@ -346,9 +362,9 @@ public class OverviewCtrl implements Main.UpdatableUI {
     }
 
     public void switchToSpanish(ActionEvent actionEvent) throws BackingStoreException {
+        userConfig.setLanguageConfig("es");
         switchLocale("messages","es");
         Image image = new Image(Objects.requireNonNull(getClass().getResource("/client/misc/es_flag.png")).toExternalForm());
-        prefs.put(SELECTED_IMAGE_KEY, "/client/misc/es_flag.png");
         menuButtonView.setImage(image);
     }
 
@@ -357,7 +373,10 @@ public class OverviewCtrl implements Main.UpdatableUI {
      * want to
      */
     public void backToStartScreen() {
-        mainCtrl.showStartScreen();
+        if(admin){
+            mainCtrl.showAdminEventOverview();
+        }
+        else mainCtrl.showStartScreen();
     }
 
     /**
@@ -524,6 +543,12 @@ public class OverviewCtrl implements Main.UpdatableUI {
     }
 
     public void refresh(Event event) {
+        String lp = userConfig.getLanguageConfig();
+        if (lp.equals("en") || lp.equals("nl") || lp.equals("es")) {
+            Image image = new Image("/client/misc/" + lp +  "_flag.png");
+            menuButtonView.setImage(image);
+        }
+
         this.event = serverUtils.getEvent(event.getId());
         options.setVisible(false);
         block.setVisible(false);
@@ -629,4 +654,11 @@ public class OverviewCtrl implements Main.UpdatableUI {
         }
     }
 
+    /**
+     * marks that an admin is accessing the event overview
+     * @param b
+     */
+    public void setAdmin(boolean b) {
+        this.admin=true;
+    }
 }
